Support stack depth tracking for easier cases
This is just a good-enough implementation that will fail horribly when branching or some generated bytecode is introduced(e.g. pop2, swap).

A better implementation must figure out branch scopes and possibly try to guess which branch was taken(e.g. when value loaded from field and no instruction in between escaped and altered state)
Also need to track full stack(pop2/swap) and types as pop2 for example works differently on category 1 or 2(long/double) types

+++ src/main/include/Method.h
@@ -37,8 +37,9 @@ public:
     return parameterTypes;
   }
 
-  const size_t getParameterCount() const {
-    return parameterTypes.size();
+  const uint8_t getParameterCount() const {
+    // JVM §4.3.3: Max length 255 parameters
+    return static_cast<const uint8_t>(parameterTypes.size());
   }
 
   bool isStatic() const {
